Field size fixes.

+++ backend/users/models.py
@@ -5,8 +5,8 @@ from django.db import models
 
 class Profile(models.Model):
     employee = models.OneToOneField(User, on_delete=models.CASCADE)
-    identification_number = models.PositiveIntegerField(null=True, blank=True)
-    phone_number = models.PositiveIntegerField(null=True, blank=True)
+    identification_number = models.PositiveBigIntegerField(null=True, blank=True)
+    phone_number = models.PositiveBigIntegerField(null=True, blank=True)
     profile_picture = models.ImageField(
         upload_to="profile_pictures/%Y/%m/%d/", null=True, blank=True
     )
